+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- API/CLI
+  - Fix empty `XDG_CONFIG_HOME` not defaulting to `$HOME/.config` #474
+
 ## 2021.07.12-12.30.59
 
 - Editor