+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Editor
   - Improve hover output, both markdown and plain text forms.
+  - Allow calling thread-first/last and thread-first-all/last-all from within the list.
 
 - API/CLI
   - Fix empty `XDG_CONFIG_HOME` not defaulting to `$HOME/.config` #474